day2/pdf-search-demo/requirements.txt
langchain
langchain-openai
chromadb
pypdf
streamlit
python-dotenv
